rename package and files so that users don't have to
