+++ trr-receptorName/receptorPackage/LICENSE
