+++ trr-receptorName/receptorPackage/receptorPackage.go
@@ -1,14 +1,15 @@
 /* TODO: Name package */
-package receptorPackageName
+package receptorPackage
 
 import (
+	receptorLog "receptor/trr-receptorName/logging"
+
 	"github.com/trustero/api/go/receptor_sdk"
 	"github.com/trustero/api/go/receptor_v1"
-	"receptor/trr-receptorName/logging"
 )
 
 const (
-	receptorName = "trr-bitbucket"
+	receptorName = "trr-receptorName"
 	serviceName  = "Custom Service"
 )
 
@@ -21,21 +22,22 @@ func GetKnownServicesImpl() []string {
 }
 
 func VerifyImpl( /* TODO: Put needed receptor creds here */ field1 int, field2 string) (ok bool, err error) {
-	receptorLog.Debug("Entering ReportImpl")
+	receptorLog.Info("Entering VerifyImpl")
 	/* TODO: Implement Verify logic here */
-	receptorLog.Debug("Leaving ReportImpl")
+	receptorLog.Info("Leaving VerifyImpl")
 	return
 }
 
 func DiscoverImpl( /* TODO: Put needed receptor creds here */ field1 int, field2 string) (svcs []*receptor_v1.ServiceEntity, err error) {
+	receptorLog.Info("Entering DiscoverImpl")
 	/* TODO: Implement Discover logic here */
-	receptorLog.Debug("Leaving ReportImpl")
+	receptorLog.Info("Leaving DiscoverImpl")
 	return
 }
 
 func ReportImpl( /* TODO: Put needed receptor creds here */ field1 int, field2 string) (evidences []*receptor_sdk.Evidence, err error) {
-	receptorLog.Debug("Entering ReportImpl")
+	receptorLog.Info("Entering ReportImpl")
 	/* TODO: Implement Report logic here */
-	receptorLog.Debug("Leaving ReportImpl")
+	receptorLog.Info("Leaving ReportImpl")
 	return
 }
